Fix for fasta file bug that crashes SNPSITES
The recently 'improved' ```sed -e '2~1 s/\\./-/g'``` substitution also changed the GATK VariantsToTable ```-GT``` label and hence a later substitution that was supposed to remove this label no longer worked properly. If fix this still does not solve the issue then we have to adjust the escape character for the point in ```sed -e 's/\\.GT/\\n/g'```.